RAG_Diabetes/pyproject.toml
[project]
name = "rag-diabetes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "langchain==0.1.20",
    "langchain-community==0.0.38",
    "langchain-core==0.1.52",
    "langchain-text-splitters==0.0.2",
    "langgraph==0.0.38",
    "google-genai>=1.48.0",
    "google-generativeai>=0.8.5",
    "chromadb>=1.3.0",
    "scikit-learn>=1.7.2",
    "torch==2.2.2",
    "pandas",
    "lightgbm",
    "numpy<2.0", 
    "streamlit",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]
